VersionedFile#finalize with empty pipeline
VersionedFile#finalize will now work when the :overwrite option is true and the file is not changed

This used to fail because if the versioned file is not changed, it would remove the original before trying to copy it

+++ lib/file_pipeline/versioned_file.rb
@@ -144,6 +144,8 @@ module FilePipeline
     #   * +true+ - The finalized version will replace the #original.
     def finalize(overwrite: false)
       yield(self) if block_given?
+      return original unless changed?
+
       filename = overwrite ? replacing_trarget : preserving_taget
       FileUtils.rm original if overwrite
       @original = Versions.copy(current, original_dir, filename)
